pyproject.toml
[tool.poetry]
name = "python learn"
version = "0.1.0"
description = ""
authors = ["samuelindrawijaya <samuelindraw75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.6.0"
flasgger = "^0.9.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
